cloudjudge changes

+++ clouddetect-persistence/src/main/java/persistence/dao/FeedDao.java
@@ -10,7 +10,7 @@ public interface FeedDao {
     
     public List<Feed> getAllActiveFeeds();
     
-    public Feed getFeed(int id);
+    public Feed getFeed(long id);
     
     public void saveResult(Feed feed);
